pyproject.toml
[tool.black]
exclude = '/(\.eggs|\.git|\.mypy_cache|\.venv.*|_build|build|dist)/'
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.poetry]
name = "anime-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Damego <danyabatueff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asyncpg = "^0.27.0"
python-dotenv = "^0.21.1"
selenium = "^4.8.0"
webdriver-manager = "^3.8.5"
aiogram = {version = "^3.0.0b6", allow-prereleases = true}
anilibria-py = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
